requirements.txt
django-robots
Markdown
django-markdown-deux
django-markdownx
